Replace custom parsing with PEG grammar and parsimonious parser compiler (#1)
Replace the custom Kusto parsing with the parsimonious parser compiler and the Kusto PEG grammar.

The downside is that I now have a package dependency, parsimonious, which people will have to install with pip. (it is not included by default in anaconda).

Another downside is that the parsing takes a bit longer than it did before. This is not noticeable when it succeeds, but if there is a syntax error it can take a half a second to explore all the possibilities before failing. I don't think this is a problem for every-day usage. However running the hundreds of unit tests can take a while if a bunch of them fail.

The upside is that the parsing is now much, much more robust and flexible. In particular, I have been able to introduce context into the syntax. For example the "by" operator is only allowed in certain tabular operators. Similarly the "*" operator is only allowed in very specific circumstances, e.g. Any(*) or distinct *

I have also been able to add support for full Kusto queries including chaining tabular operators together with pipes, and even joining multiple query statements together with ;.